requirements.txt
dspy-ai
pandas
matplotlib
pyarrow
